Add weighting to answers

+++ code/kuusi/web/rest/facetteselection.py
@@ -79,7 +79,7 @@ class FacetteSelectionViewSet(ListModelMixin, GenericViewSet, DestroyModelMixin)
         },
         parameters=[ 
           OpenApiParameter("session_pk", OpenApiTypes.STR, OpenApiParameter.PATH, required=True),
-          OpenApiParameter("reset", OpenApiTypes.BOOL, OpenApiParameter.QUERY, required=False, description="Force existing selections on the same topic to be deleted. Used for facette selections bound to radio selections"),
+          OpenApiParameter("reset", OpenApiTypes.STR, OpenApiParameter.QUERY, required=False, description="Force existing selections on the same topic to be deleted. Used for facette selections bound to radio selections. Values: 'all', 'this'"),
         ],
     )
     def create(self, request, session_pk) -> FacetteSelection:
@@ -89,10 +89,13 @@ class FacetteSelectionViewSet(ListModelMixin, GenericViewSet, DestroyModelMixin)
         if session is None or facette is None:
             return Response(status=status.HTTP_404_NOT_FOUND)
         
-        is_reset = str(request.query_params.get('reset')).lower() == "true"
+        is_reset = str(request.query_params.get('reset')).lower()
         if is_reset:
-            FacetteSelection.objects.filter(session=session).filter(facette__topic=facette.topic).delete()
-        
+            if is_reset == "all":
+                FacetteSelection.objects.filter(session=session).filter(facette__topic=facette.topic).delete()
+            if is_reset == "this":
+                FacetteSelection.objects.filter(session=session).filter(facette=facette).delete()
+    
         selections = FacetteSelection.objects.filter(session=session).filter(facette=facette)
         if selections.count() != 0:
             return Response(status=status.HTTP_409_CONFLICT)
